+++ Adj/Program/Datastructure/Implementation/List.hs
@@ -2,6 +2,6 @@ module Adj.Program.Datastructure.Implementation.List where
 
 import Adj.Auxiliary (type (=!?=))
 import Adj.Program.Generation (Construction)
-import Adj.Program.Option (Option)
+import Adj.Program.Maybe (Maybe)
 
-type List = Option =!?= Construction Option
+type List = Maybe =!?= Construction Maybe